Disclaim that iOS development needs the latest Xcode.
While OS X 10.7 is fine for Meteor itself, iOS development (via Cordova)
almost always requires a recent, or the latest version of Xcode.

Xcode itself generally has the requirement of the most recent version of
OS X/macOS so it may be that a developer needs the most recent OS X
version in order to do iOS development, but the rest of Meteor should
work just fine.

Helps with meteor/meteor#7210.